chore(astronvim): move 'autocmds.lua' to 'extras/autocmds.lua'
  ## what
  - Move dotfiles/astronvim/autocmds.lua to dotfiles/tools/astronvim/extras/autocmds.lua
  - Update autocmds import in dotfiles/astronvim/polish.lua

  ## how

  ## why
  - to manage files
  - to keep Astronvim user config folder clean
  - extra/ folder can be used for adding functionality thats not
    directly supported by Astronvim

  ## where

  ## usage
