+++ tools/astronvim/polish.lua
@@ -1,5 +1,5 @@
 -- This function is run last
 -- good place to configure mappings and vim options
 return function()
-  require("user.autocmds")
+  require("user.extras.autocmds")
 end
